feat: permisstion

+++ database/migrations/2026_04_13_124121_add_is_issued_to_invoices_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (! Schema::hasTable('invoices')) {
+            return;
+        }
+
         Schema::table('invoices', function (Blueprint $table) {
             $table->boolean('is_issued')->default(false)->after('is_paid')->comment('Đã xuất hóa đơn hay chưa');
         });
@@ -21,6 +25,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (! Schema::hasTable('invoices')) {
+            return;
+        }
+
         Schema::table('invoices', function (Blueprint $table) {
             $table->dropColumn('is_issued');
         });
